[kotlin] Dont rely on global useCompositeAnalysis for resolution
A couple of previous commits, as well as fixes to KTIJ-20716 introduced
an ability to analyze some code in a separate platform-specific facade
(as if the SEPARATE mode is enabled), even if globally COMPOSITE mode is
enabled.

Therefore, all remnants of checking project.useCompositeAnalysis were
removed from the resolution-setup infrastructure (as whether we should
use COMPOSITE or SEPARATE mode might be different depending on the
specific facade).

Instead, checking instance of AnalaysisSettings should be used. This
helps resolution infra to make a consistent choices regarding analysis
mode (for example, before this commit we were injecting KotlinTypeRefinerImpl
even in platform-specific facades, which was leading to weird failures
and exceptions, e.g. ReenterantLazyValueComputationException and such)

Note that, as consequence, we can't decide on useTypeRefinement
analysis flag for LanguageVersionSettings, because one module might be
analyzed in different facades (and thus have different values of that
flag). This commit removes setting up this flag, so it should be used
together with corresponding commit on kotlin.git side, which removes
checking of that AnalysisFlag from frontend and migrated to checking
refiner type

^KTIJ-20775

GitOrigin-RevId: 9f51ec4972dcaab9601d50d3da03e840c5904dd5